+++ serialization/common/common/include/ecal/msg/exception.h
@@ -20,6 +20,7 @@
 #pragma once
 
 #include <exception>
+#include <string>
 
 namespace eCAL
 {
