+++ serialization/protobuf/protobuf/include/ecal/msg/protobuf/dynamic_json_subscriber.h
@@ -24,160 +24,11 @@
 
 #pragma once
 
-#include <ecal/ecal.h>
-#include <ecal/os.h>
-#include <ecal/msg/exception.h>
 #include <ecal/msg/subscriber.h>
-#include <ecal/msg/protobuf/ecal_proto_dyn.h>
-
-#include <iostream>
-#include <sstream>
-#include <fstream>
-#include <memory>
-#include <cstdio>
-#include <string>
-
-#ifdef _MSC_VER
-#pragma warning(push, 0) // disable proto warnings
-#endif
-#ifdef __GNUC__
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wunused-parameter"
-#endif
-#include <google/protobuf/stubs/common.h>
-#include <google/protobuf/util/json_util.h>
-#include <google/protobuf/util/type_resolver_util.h>
-#include <google/protobuf/util/type_resolver.h>
-#ifdef _MSC_VER
-#pragma warning(pop)
-#endif
-#ifdef __GNUC__
-#pragma GCC diagnostic pop
-#endif
-
-namespace
-{
-  /* @cond */
-  inline bool StrEmptyOrNull(const std::string& str)
-  {
-    if (str.empty())
-    {
-      return true;
-    }
-    else
-    {
-      for (auto c : str)
-      {
-        if (c != '\0')
-        {
-          return false;
-        }
-      }
-      return true;
-    }
-  }
-  /* @endcond */
-}
+#include <ecal/msg/protobuf/dynamic_serializer.h>
 
 namespace eCAL
 {
-    namespace internal
-    {
-      class ProtobufDynamicJSONDeserializer
-      {
-      public:
-        static SDataTypeInformation GetDataTypeInformation()
-        {
-          SDataTypeInformation topic_info;
-          topic_info.encoding = "proto";
-          topic_info.name = "*";
-          topic_info.descriptor = "*";
-          return topic_info;
-        }
-
-        std::string Deserialize(const void* buffer_, size_t size_, const SDataTypeInformation& datatype_info_)
-        {
-          google::protobuf::util::JsonPrintOptions options;
-#if GOOGLE_PROTOBUF_VERSION >= 5026000
-          options.always_print_fields_with_no_presence = true;
-#else
-          options.always_print_primitive_fields = true;
-#endif
-
-          std::string binary_input;
-          binary_input.assign(static_cast<const char*>(buffer_), static_cast<size_t>(size_));
-          std::string json_output;
-          auto status = google::protobuf::util::BinaryToJsonString(GetTypeResolver(datatype_info_).get(), GetQualifiedTopicType(datatype_info_), binary_input, &json_output, options);
-          if (status.ok())
-          {
-            return json_output;
-          }
-          else
-          {
-            throw DeserializationException("Error deserializing Protobuf data to json object.");
-          }
-        }
-
-      private:
-        std::shared_ptr<google::protobuf::util::TypeResolver> GetTypeResolver(const SDataTypeInformation& datatype_info_)
-        {
-          auto schema = m_type_resolver_map.find(datatype_info_);
-          if (schema == m_type_resolver_map.end())
-          {
-            m_type_resolver_map[datatype_info_] = CreateTypeResolver(datatype_info_);
-          }
-          return m_type_resolver_map[datatype_info_];
-        }
-
-        std::shared_ptr<google::protobuf::util::TypeResolver> CreateTypeResolver(const SDataTypeInformation& datatype_info_)
-        {
-          std::string unqualified_topic_type = GetUnqualifiedTopicType(datatype_info_);
-
-          if (StrEmptyOrNull(unqualified_topic_type))
-          {
-            throw DeserializationException("ProtobufDynamicJSONDeserializer: Could not get type");
-          }
-
-          std::string topic_desc = datatype_info_.descriptor;
-          if (StrEmptyOrNull(topic_desc))
-          {
-            throw DeserializationException("ProtobufDynamicJSONDeserializer: Could not get description for type" + std::string(unqualified_topic_type));
-          }
-
-          google::protobuf::FileDescriptorSet proto_desc;
-          proto_desc.ParseFromString(topic_desc);
-          std::string error_s;
-          const std::shared_ptr<google::protobuf::Message> msg(m_dynamic_decoder.GetProtoMessageFromDescriptorSet(proto_desc, unqualified_topic_type, error_s));
-          std::shared_ptr<google::protobuf::util::TypeResolver> resolver{ google::protobuf::util::NewTypeResolverForDescriptorPool("", m_dynamic_decoder.GetDescriptorPool()) };
-
-          if (resolver == nullptr)
-          {
-            std::stringstream s;
-            s << "ProtobufDynamicJSONDeserializer: Message of type " + unqualified_topic_type << " could not be decoded" << std::endl;
-            s << error_s;
-            throw DeserializationException(s.str());
-          }
-
-          return resolver;
-        }
-
-        std::string GetQualifiedTopicType(const SDataTypeInformation& data_type_info_)
-        {
-            return  "/" + data_type_info_.name;
-        }
-
-        std::string GetUnqualifiedTopicType(const SDataTypeInformation& data_type_info_)
-        {
-          const auto& type_name = data_type_info_.name;
-          return  type_name.substr(type_name.find_last_of('.') + 1, type_name.size());
-        }
-
-
-        eCAL::protobuf::CProtoDynDecoder                                                      m_dynamic_decoder;
-        std::map<SDataTypeInformation, std::shared_ptr<google::protobuf::util::TypeResolver>> m_type_resolver_map;
-      };
-    }
-
     namespace protobuf
     {
       /**
